fix(qa): preserve packaged auth redaction boundary
